+++ python/cudf/dataframe/series.py
@@ -77,9 +77,7 @@ class Series(object):
         if not isinstance(data, columnops.TypedColumnBase):
             data = columnops.as_column(data, nan_as_null=nan_as_null)
 
-        if isinstance(index, range):
-            index = RangeIndex(index)
-        if index is not None and not isinstance(index, (RangeIndex, Index,)):
+        if index is not None and not isinstance(index, Index):
             raise TypeError('index not a Index type: got {!r}'.format(index))
 
         assert isinstance(data, columnops.TypedColumnBase)
@@ -301,8 +299,11 @@ class Series(object):
         # Prepare cells
         cols = OrderedDict([('', self.values_to_string(nrows=nrows))])
         # Format into a table
-        return formatting.format(index=self.index,
-                                 cols=cols, more_rows=more_rows)
+        output = formatting.format(index=self.index,
+                                   cols=cols, more_rows=more_rows,
+                                   series_spacing=True)
+        return output + "\nName: {}, dtype: {}".format(self.name, self.dtype)\
+            if self.name else output + "\ndtype: {}".format(self.dtype)
 
     def __str__(self):
         return self.to_string(nrows=10)
@@ -1037,6 +1038,32 @@ class Series(object):
 
         return Series(numerical.column_hash_values(self._column))
 
+    def hash_encode(self, stop, use_name=False):
+        """Encode column values as ints in [0, stop) using hash function.
+
+        Parameters
+        ----------
+        stop : int
+            The upper bound on the encoding range.
+        use_name : bool
+            If ``True`` then combine hashed column values
+            with hashed column name. This is useful for when the same
+            values in different columns should be encoded
+            with different hashed values.
+        Returns
+        -------
+        result: Series
+            The encoded Series.
+        """
+        assert stop > 0
+
+        from . import numerical
+        initial_hash = np.asarray(hash(self.name)) if use_name else None
+        hashed_values = numerical.column_hash_values(
+            self._column, initial_hash_values=initial_hash)
+        hashed_values = np.mod(hashed_values, stop)
+        return Series(hashed_values)
+
     def quantile(self, q, interpolation='midpoint', exact=True,
                  quant_index=True):
         """
@@ -1068,6 +1095,10 @@ class Series(object):
         else:
             return Series(self._column.quantile(q, interpolation, exact),
                           index=as_index(np.asarray(q)))
+
+    def groupby(self, group_series):
+        from cudf.groupby.groupby import SeriesGroupBy
+        return SeriesGroupBy(self, group_series)
 
 
 register_distributed_serializer(Series)